A relatively large set of small changes to make everything compile without warnings using "-W -Wall -Wno-unused -pedantic". Mainly sign-correctness on integers and some const correctness tidy-ups. The big deal is the change of function pointers to have void(foo*)() as the "generic" type to store in the callback table as opposed to the illegal void*. This change entailed a few modifications in plugin_api.* and in the actual plugin code, but it should be fairly straight-forward.

+++ src/blocks/plant.cpp
@@ -103,7 +103,7 @@ void BlockPlant::remBlock(PlantBlock* p2)
 
 void BlockPlant::addBlock(PlantBlock* p2)
 {
-  for (int i = 0; i < growingPlants.size(); i++)
+  for (size_t i = 0; i < growingPlants.size(); i++)
   {
     PlantBlock* p = growingPlants[i];
     if (p->x == p2->x && p->y == p2->y && p->z == p2->z && p->map == p2->map)
